scripts/pyproject.toml
[project]
name = "scripts-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jinja2==3.1.2",
    "psycopg2-binary==2.9.6",
    "python-dotenv==1.0.0",
    "pytz==2023.3",
    "requests==2.28.2",
    "requests-oauthlib==1.3.1",
    "schedule==1.2.0",
]
